Move endpointmanager code to cmd/sendendpoints
